+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 # All changes in project EOCS
 
+## [0.0.5] - 2026.04.17
+
+### Added
+- Text Render (Padding, Resazing, Cordinates)
+- [Vertex Shader](Assets/shaders/shader.vert) and [Fragment Shader](Assets/shaders/shader.frag) from text
+
 ## [0.0.4] - 2026.04.17
 
 ### Added
@@ -10,8 +16,8 @@
 ## [0.0.3] - 2026.04.17
 
 ### Changed
-- Renamed project files and directory from '3dGame' to 'EOCS'
-- Updated root namespace from 'BPX' to 'EOCS'
+- Renamed project files and directory from **3dGame** to **EOCS**
+- Updated root namespace from **BPX** to **EOCS**
 - README.md